requirements.txt
yahoo-finance
django
bitcoin-price-api
forex-python
djangorestframework